edit paciente?

+++ api/src/controllers/paciente-controller.js
@@ -1,3 +1,4 @@
+const { AgendamentoModel } = require('../models/agendamento-model');
 const { PacienteModel } = require('../models/paciente-model');
 const paginate = require('../utils/paginacao');
 const { Op } = require('sequelize');
@@ -51,6 +52,14 @@ class PacienteController {
   async delete(req, res) {
     try {
       const { id } = req.params;
+
+      // Verifique se existem agendamentos para o paciente
+      const pacienteAgendamentos = await AgendamentoModel.findOne({ where: { id_paciente: id } });
+
+      if (pacienteAgendamentos) {
+        return res.status(400).json({ error: 'Paciente tem agendamentos' });
+      }
+
       await PacienteModel.destroy({ where: { id_paciente: id } });
       return res.status(204).send();
     } catch (error) {
@@ -59,33 +68,33 @@ class PacienteController {
     }
   }
 
-    async getAll(req, res) {
-      try {
-        const { page } = req.params;
-        const { nome, limit } = req.query; // Parâmetros da consulta
-        let pacientes; // Variável para armazenar os pacientes
-    
-        // Opções de filtro para o Sequelize
-        const filterOptions = {};
-    
-        if (nome) {
-          // Adicione a opção de filtro por nome, ignorando acentuações
-          filterOptions.where = {
-            nome: {
-              [Op.iLike]: `%${nome}%`
-            }
-          };
-        }
-    
-        // Chama a função paginate com as opções de filtro
-        pacientes = await paginate(PacienteModel, page, limit, filterOptions);
-    
-        return res.status(200).json(pacientes); // Retorna os pacientes encontrados e paginados
-      } catch (error) {
-        console.error(error);
-        return res.status(500).json({ error: 'Internal server error' });
+  async getAll(req, res) {
+    try {
+      const { page } = req.params;
+      const { nome, limit } = req.query; // Parâmetros da consulta
+      let pacientes; // Variável para armazenar os pacientes
+
+      // Opções de filtro para o Sequelize
+      const filterOptions = {};
+
+      if (nome) {
+        // Adicione a opção de filtro por nome, ignorando acentuações
+        filterOptions.where = {
+          nome: {
+            [Op.iLike]: `%${nome}%`
+          }
+        };
       }
+
+      // Chama a função paginate com as opções de filtro
+      pacientes = await paginate(PacienteModel, page, limit, filterOptions);
+
+      return res.status(200).json(pacientes); // Retorna os pacientes encontrados e paginados
+    } catch (error) {
+      console.error(error);
+      return res.status(500).json({ error: 'Internal server error' });
     }
   }
+}
 
 module.exports = new PacienteController()
